+++ NEWS.md
@@ -2,7 +2,10 @@
 
 * API breaking: changed "dependencies" to "depend_sp" in `indirect_relations()`
 * API breaking: changed "geodesic" to "dist_sp" in `indirect_relations()`
-* added indirect relations: log_forest,depend_netflow, depend_exp
+* API breaking: changed "resistance" to "dist_resist" in `indirect_relations()`
+* The above old types still work
+* added indirect relations: dist_lf,depend_netflow, 
+depend_exp, depend_rsps, depend_rspn, depend_curflow
 * changed `require` to `library` in examples
 
 ## netrankr 0.1.1