+++ NEWS.md
@@ -1,6 +1,6 @@
 ## netrankr 0.2.0.9000
 
-* added Rstudio addin to build centrality indices
+* added Rstudio addin to build more than 20 centrality indices
 * added indirect relations: dist_lf,dist_walk, depend_netflow, 
 depend_exp, depend_rsps, depend_rspn, depend_curflow,dist_rwalk
 * API breaking: changed "dependencies" to "depend_sp" in `indirect_relations()`